Add after suite to tests, refactor file directories
