+++ api/api.go
@@ -8,7 +8,7 @@ import (
   "encoding/json"
   "database/sql"
 
-  "github.com/anon0mys/qs_golang/models"
+  "github.com/anon0mys/qs_golang/internal/models"
 
   "github.com/gorilla/mux"
   _ "github.com/lib/pq"
@@ -17,6 +17,7 @@ import (
 type App struct {
   Router *mux.Router
   DB     *sql.DB
+  Server *http.Server
 }
 
 func (a *App) Initialize(user, password, dbname string) {
@@ -30,15 +31,29 @@ func (a *App) Initialize(user, password, dbname string) {
   }
 
   a.Router = mux.NewRouter()
+  a.Server = &http.Server{Addr: ":3000", Handler: a.Router}
   a.initializeRoutes()
 }
 
-func (a *App) Run(addr string) {
-  log.Fatal(http.ListenAndServe(":3000", a.Router))
+func (a *App) Run() {
+  log.Fatal(a.Server.ListenAndServe())
 }
 
 func (a *App) initializeRoutes() {
   a.Router.HandleFunc("/api/v1/foods", a.CreateFood).Methods("POST")
+  a.Router.HandleFunc("/api/v1/foods", a.CreateFood).Methods("GET")
+}
+
+func respondWithError(w http.ResponseWriter, code int, message string) {
+  respondWithJSON(w, code, map[string]string{"error": message})
+}
+
+func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
+  response, _ := json.Marshal(payload)
+
+  w.Header().Set("Content-Type", "application/json")
+  w.WriteHeader(code)
+  w.Write(response)
 }
 
 func (a *App) CreateFood(w http.ResponseWriter, r *http.Request) {
@@ -58,14 +73,6 @@ func (a *App) CreateFood(w http.ResponseWriter, r *http.Request) {
   respondWithJSON(w, http.StatusCreated, f)
 }
 
-func respondWithError(w http.ResponseWriter, code int, message string) {
-  respondWithJSON(w, code, map[string]string{"error": message})
-}
+func (a *App) GetFood(w http.ResponseWriter, r *http.Request) {
 
-func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
-  response, _ := json.Marshal(payload)
-
-  w.Header().Set("Content-Type", "application/json")
-  w.WriteHeader(code)
-  w.Write(response)
 }
