+++ api/api_test.go
@@ -25,7 +25,7 @@ var _ = BeforeSuite(func() {
     os.Getenv("QS_GOLANG_DB_PASSWORD"),
     os.Getenv("QS_GOLANG_DB_NAME"))
 
-  app.Run(":3000")
+  app.Run()
 
   if _, err := app.DB.Exec(tableCreationQuery); err != nil {
       log.Fatal(err)
@@ -33,6 +33,10 @@ var _ = BeforeSuite(func() {
 
   app.DB.Exec("DELETE FROM foods")
   app.DB.Exec("ALTER SEQUENCE foods_id_seq RESTART WITH 1")
+})
+
+var _ = AfterSuite(func() {
+  os.Exit(0)
 })
 
 const tableCreationQuery = `CREATE TABLE IF NOT EXISTS foods
