+++ api/food_enpoints_test.go
@@ -13,7 +13,7 @@ import (
 var _ = Describe("Foods API", func() {
 	Describe("POST /api/v1/foods", func() {
 		Context("with valid parameters", func() {
-			It("should return the created food", func() {
+			It("should return the created food", func(done Done) {
 				foodParams := []byte(`{"name":"Banana","calories":100}`)
 
 				req, _ := http.NewRequest("POST", "/api/v1/foods", bytes.NewBuffer(foodParams))
